*Updates*
Features Added:
1.) Added a step view which guides users through each step of the appointment process.
2.) Users are able to select a date and time and store their selection into FireStore database, which prevents other users selecting their time if they booked it.
3.) Change the layout design for register page and login page.

Issues trying to solve:

1.) In the schedule section, I am trying to get one of the radio button text to be displayed
at the end of the confirmation page. I have tried storing the text to firestore database and
then retrieving, but that doesn't work. Because I am using view pager which shares all screens so it
can't display the radioButton text.

2.) At the date/time section in Schedule, I tried to disable certain dates for that calendar like
weekends only and I can't seem to disable it. The source for this calendar I used is: https://github.com/Mulham-Raee/Horizontal-Calendar.

3.) Lastly, storing user booking information to fireStore database. I tried to initiate user information as soon as they log in so they I can invoke their name, address, and zip code, in order for it to be stored separately in their Firebase booked appointment section. I get an error stating that it's a null object reference meaning I haven't set a value to it.